+++ src/main/java/stevekung/mods/moreplanets/planets/fronos/blocks/fluids/BlockFluidMineralWater.java
@@ -39,7 +39,7 @@ public class BlockFluidMineralWater extends BlockFluidBaseMP
     @SideOnly(Side.CLIENT)
     public void randomDisplayTick(World world, BlockPos pos, IBlockState state, Random rand)
     {
-        int meta = (Integer) state.getValue(LEVEL);
+        int meta = state.getValue(LEVEL);
 
         if (rand.nextInt(1) == 0)
         {
