+++ src/main/java/stevekung/mods/moreplanets/planets/fronos/entities/ai/EntityAIFronosVillagerHarvestFarmland.java
@@ -80,7 +80,7 @@ public class EntityAIFronosVillagerHarvestFarmland extends EntityAIMoveToBlock
             IBlockState iblockstate = world.getBlockState(blockpos);
             Block block = iblockstate.getBlock();
 
-            if (this.field_179501_f == 0 && block instanceof BlockCrops && ((Integer)iblockstate.getValue(BlockCrops.AGE)).intValue() == 7)
+            if (this.field_179501_f == 0 && block instanceof BlockCrops && iblockstate.getValue(BlockCrops.AGE).intValue() == 7)
             {
                 world.destroyBlock(blockpos, true);
             }
@@ -140,7 +140,7 @@ public class EntityAIFronosVillagerHarvestFarmland extends EntityAIMoveToBlock
             IBlockState iblockstate = world.getBlockState(pos);
             block = iblockstate.getBlock();
 
-            if (block instanceof BlockCrops && ((Integer)iblockstate.getValue(BlockCrops.AGE)).intValue() == 7 && this.field_179503_e && (this.field_179501_f == 0 || this.field_179501_f < 0))
+            if (block instanceof BlockCrops && iblockstate.getValue(BlockCrops.AGE).intValue() == 7 && this.field_179503_e && (this.field_179501_f == 0 || this.field_179501_f < 0))
             {
                 this.field_179501_f = 0;
                 return true;
